Screen fading script for X
This is a Ruby port of the Bash script at:
http://console.systems/2016/10/handmade-screen-locker-with-fade-out.html

The only major change is to not sleep for a fixed amount at each step,
but instead to sleep until the step should be over (or don't sleep if
we're behind).

Unfortunately, the implementation being based on running xcalib in a
sub-shell seems to mean that each time step (for any reasonable number
of steps) takes too long.