Ema uses Emacs frame in current i3 workspace
If you have one Emacs server with frames that are present on multiple
workspaces in i3, Emacs doesn't have any way of determining which ones
are visible or not. Instead, it just uses the most recently focused
frame, even if that is in a different workspace.

This change uses i3-msg to attempt to focus an Emacs window that is in
the currently focused workspace. If it succeeds we use that frame for
the client command (implicitly because it was just focused), otherwise
we create a new frame.